添加了图片的缩略图剪辑功能，为carousel做准备

+++ jobs/remove_redundant_image_job.js
@@ -5,14 +5,9 @@
 let database = require('../model').loader;
 let path = require('path');
 let fs = require('fs');
+let image_tools = require('../libs/image');
 
 module.exports = function (callback) {
-
-    function get_thumb_path(image_path) {
-        let dirname = path.dirname(image_path);
-        let extname = path.extname(image_path);
-        return path.join(dirname, path.basename(image_path, extname) + "_thumb" + extname);
-    }
 
     let Diary = database.models.diary;
     let Photo = database.models.photo;
@@ -23,14 +18,15 @@ module.exports = function (callback) {
         if (err) callback(err);
         photos.forEach(function (photo) {
             image_ids.push(photo.path);
-            image_ids.push(get_thumb_path(photo.path));
+            image_ids.push(image_tools.get_thumb_path(photo.path));
+            image_ids.push(image_tools.get_thumb_path(photo.path));
         });
         Diary.find({}, function(err, diaries) {
             if (err) callback(err);
             diaries.forEach(function (diary) {
                 for(let i = 0; i < diary.images.length; i++) {
                     image_ids.push(diary.images[i]);
-                    image_ids.push(get_thumb_path(diary.images[i]));
+                    image_ids.push(image_tools.get_thumb_path(diary.images[i]));
                 }
             });
             Guest.find({}, function (err, guests) {
@@ -51,12 +47,14 @@ module.exports = function (callback) {
                     let total_number = Object.keys(delete_set).length;
                     let counter = 0;
                     for(let key in delete_set) {
-                        fs.unlink(path.join(image_path, key), function () {
-                            counter++;
-                            if (counter >= total_number) {
-                                callback();
-                            }
-                        });
+                        if (fs.existsSync(path.join(image_path, key))) {
+                            fs.unlink(path.join(image_path, key), function () {
+                                counter++;
+                                if (counter >= total_number) {
+                                    callback();
+                                }
+                            });
+                        }
                     }
                 });
             });
